+++ packages/angular/src/icon-content-switcher/index.ts
@@ -1,0 +1,3 @@
+export { IconContentSwitcherOption } from './icon-content-switcher-option.directive';
+export { IconContentSwitcher } from './icon-content-switcher.component';
+export { IconContentSwitcherModule } from './icon-content-switcher.module';