+++ packages/react/src/components/DateTimePicker/DateTimePickerV2.test.e2e.jsx
@@ -47,6 +47,8 @@ describe('DateTimePickerV2', () => {
             start: Cypress.sinon.match.any,
             startDate: '08/06/2021',
             startTime: '01:00',
+            humanValue: '2021-08-06 01:00 to 2021-08-08 12:34',
+            tooltipValue: '',
           },
         });
       });
@@ -175,6 +177,8 @@ describe('DateTimePickerV2', () => {
           timeRangeValue: {
             relativeToWhen: '',
             relativeToTime: '12:04',
+            humanValue: '',
+            tooltipValue: '',
           },
         });
       });
@@ -212,6 +216,10 @@ describe('DateTimePickerV2', () => {
             relativeToTime: '12:04',
             end: Cypress.sinon.match.any,
             start: Cypress.sinon.match.any,
+            humanValue: Cypress.sinon.match((value) => {
+              return value.includes('Invalid Date') && value.includes('12:04');
+            }),
+            tooltipValue: '',
           },
         });
       });