+++ packages/angular/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.147.0-next.22](https://github.com/IBM/carbon-components-angular/compare/v2.147.0-next.21...v2.147.0-next.22) (2021-10-21)
+
+
+### Bug Fixes
+
+* **build:** move entry points to the same level ([4a4756b](https://github.com/IBM/carbon-components-angular/commit/4a4756b1de2a4e4cdb82cd12b1d197ce1c45fa16))
+* **list-header:** change imports ([6fcf26e](https://github.com/IBM/carbon-components-angular/commit/6fcf26ea86cd549ceebb6d59cf4e57e25aeaa1f0))
+
+
+
+
+
+# [2.147.0-next.12](https://github.com/IBM/carbon-components-angular/compare/v2.147.0-next.11...v2.147.0-next.12) (2021-10-06)
+
+
+### Features
+
+* **icon-content-switcher:** add it ([310aa15](https://github.com/IBM/carbon-components-angular/commit/310aa15c4214ec6745659b138286389c1338fd12))
+
+
+
+
+
+# [2.147.0-next.11](https://github.com/IBM/carbon-components-angular/compare/v2.147.0-next.10...v2.147.0-next.11) (2021-10-06)
+
+
+### Bug Fixes
+
+* **table:** body ([f1cd8d6](https://github.com/IBM/carbon-components-angular/commit/f1cd8d677f7bb0afa20b448223484314d3300e05))
+* **table:** projectedRowLength ([f9740b7](https://github.com/IBM/carbon-components-angular/commit/f9740b76cb3194e8af2c32de7059df61e489bbd8))
+* **table:** projectedRowLength now does rowSpans ([0312581](https://github.com/IBM/carbon-components-angular/commit/031258157e03ec4a01ecf00a7b2168463cc3b703))
+
+
+
+
+
 # [2.147.0-next.0](https://github.com/IBM/carbon-components-angular/compare/v2.146.2-next.2...v2.147.0-next.0) (2021-09-27)